Rename equity minute writer module.
Prepare for reader in same module.
